[front] - refactor: Extract PlotBlock to separate file
 - Move PlotBlock component from RenderMessageMarkdown.tsx to RenderCodePlot.tsx
 - Update import statements in AgentMessage.tsx
