+++ front/components/assistant/conversation/AgentMessage.tsx
@@ -38,7 +38,8 @@ import {
   providerFromDocument,
   titleFromDocument,
 } from "@app/components/assistant/conversation/RetrievalAction";
-import { RenderMessageMarkdown, PlotBlock } from "@app/components/assistant/RenderMessageMarkdown";
+import { RenderMessageMarkdown } from "@app/components/assistant/RenderMessageMarkdown";
+import { PlotBlock} from "@app/components/assistant/RenderCodePlot";
 import { useEventSource } from "@app/hooks/useEventSource";
 import { useSubmitFunction } from "@app/lib/client/utils";
